requirements.txt
streamlit
langchain-core
langchain-openai
fpdf
python-dotenv
